Change type double to BigDecimal

+++ src/main/java/natasha/shop/Product.java
@@ -1,14 +1,16 @@
 package natasha.shop;
 
+import java.math.BigDecimal;
+
 public class Product {
     private Category category;
     private String name;
-    private double price;
+    private BigDecimal price;
 
     public Product() {
     }
 
-    public Product(Category category, String name, double price) {
+    public Product(Category category, String name, BigDecimal price) {
         this.category = category;
         this.name = name;
         this.price = price;
@@ -30,11 +32,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
